requirements.txt
sentence-transformers==2.2.2
huggingface_hub==0.14.1
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain-google-genai
langchain_experimental
pandas
openpyxl
pypdf